minor build warning fix
When you use new[] to allocate an array of objects, you must use delete[] to deallocate that memory. Using delete without the brackets is incorrect and can lead to undefined behavior, including memory leaks or crashes, because delete doesn't know it's dealing with an array and won't properly call destructors for each element in the array.