title id lookup now handled locally
Since the output from marketplace.xbox.com can no longer be parsed to get the titleID, I've made it local, in the form of gamedatabase.json. This is a rough implementation, to keep compatibility with calling functions, and minimal code change. Title ID Finder and Gamer Picture Pack Creator should be functional again.
json file is in the project folder.